Subnet builder tests
