+++ src/main/python/powerFactory2json/pf2jsonUtils.py
@@ -25,6 +25,7 @@ attributes4export = {
     'cpArea': [],
     'nodes': [
         "vtarget",
+        "uknom",
         "cpZone",
         "cpArea",
         "GPSlat",
