+++ src/main/scala/edu/ie3/powerFactory2psdm/converter/GridConverter.scala
@@ -12,15 +12,23 @@ import edu.ie3.powerFactory2psdm.model.powerfactory.{
 }
 import edu.ie3.powerFactory2psdm.util.GridPreparator
 
+/**
+  * Functionalities to transform an exported and then parsed PowerFactory grid to the PSDM.
+  */
 case object GridConverter {
 
   def convert(pfGrid: PowerFactoryGrid) = {
     val gridElements = convertGridElements(pfGrid)
   }
 
+  /**
+    * Converts the grid elements of the PowerFactory grid
+    * @param rawPfGrid the raw parsed PowerFactoryGrid
+    */
   def convertGridElements(rawPfGrid: PowerFactoryGrid): Unit = {
     val pfGrid = GridPreparator.prepare(rawPfGrid)
     val pfGridMaps = new PowerFactoryGridMaps(pfGrid)
     val graph = GridGraphBuilder.build(pfGridMaps)
+    val subnets = SubnetBuilder.buildSubnets(graph, pfGridMaps.UUID2node)
   }
 }
