+++ src/main/scala/edu/ie3/powerFactory2psdm/exception/pf/GridConfigurationException.scala
@@ -1,0 +1,12 @@
+/*
+ * © 2021. TU Dortmund University,
+ * Institute of Energy Systems, Energy Efficiency and Energy Economics,
+ * Research group Distribution grid planning and operation
+ */
+
+package edu.ie3.powerFactory2psdm.exception.pf
+
+case class GridConfigurationException(
+    private val msg: String,
+    private val cause: Throwable = None.orNull
+) extends PfException(msg, cause)
