+++ src/main/scala/edu/ie3/powerFactory2psdm/model/powerfactory/PowerFactoryGrid.scala
@@ -90,7 +90,8 @@ object PowerFactoryGrid {
       GPSlat: Option[Double],
       id: Option[String],
       GPSlon: Option[Double],
-      cpArea: Option[List[Option[CpArea]]]
+      cpArea: Option[List[Option[CpArea]]],
+      uknom: Option[Double]
   )
 
   final case class Trafos2w(
